Added MEF support for backends
The available backends are now loaded via MEF.  We look for backends in
the current assembly and in any DLLs that are in the same directory as
the EXE.  We could change this to look in a Plugins directory if we
thought that would be cleaner.  I went with a single directory for
simplicity.

Log4net is now initialized before the Statd instance is created so that
we can log configuration details before the statsd instance is created.