Added Example External Backend assembly
Created a File Backend that simply appends lines to a configured text
file.  This is just a proof of concept that backends can be defined and
loaded from an external dll.  The project currently references a new
project - statsd.net.core.  This assembly has just the basics needed for
a backend - the IBackend interface and a few more classes tied to it.
The statsd.net.shared assembly references a bunch of other projects that
we wouldn't want to be required for every backend.

The example backend does not currently add itself to the main bin
directory.  You'll have to do that manually.  This could be added if we
thought it was helpful.