Consistency between GET and POST splitting
Allow for POST data to be split by comma just like the GET. Makes it
easy to switch between GET and POST while still allowing a new line to
separate stats when sending in bulk.

Also dump the failed line information to the log to assist in debugging
when you send something that fails to parse.